refactor(hydration): migrate hydration feature to shared health module
This commit moves the hydration feature from the application-specific directory to a shared `:features:health:hydration` module. This refactoring improves project structure and allows for better code reuse across different application targets.

- **`HydrationUiRoute.kt` & `HydrationScreen`**:
    - Relocated the hydration UI components to the new `com.zoewave.probase.features.health.hydration` package.
    - Updated `HydrationUiRoute` and `HydrationScreen` to use a specific `onNavigateToSettings` callback instead of a generic `navTo` navigation lambda.
- **`InputUiRoute.kt`**:
    - Updated imports to point to the new hydration feature package.
    - Adjusted the `HydrationScreen` call to provide the `onNavigateToSettings` parameter.
- **`goSwiftNavEntryProvider.kt`**:
    - Updated the navigation graph entry for the hydration destination to use the new module and updated API.
- **Build Configuration**:
    - Removed the local `:applications:goswift:apps:mobile:features:hydration` module and its associated files (resource strings, ProGuard rules, and unit tests).
    - Updated `build.gradle.kts` in both the mobile app and the input feature module to depend on the new `:features:health:hydration` library.
