+++ applications/goswift/apps/mobile/build.gradle.kts
@@ -38,7 +38,7 @@ dependencies {
     implementation(project(":applications:goswift:apps:mobile:features:home"))
     implementation(project(":applications:goswift:apps:mobile:features:shots"))
     implementation(project(":applications:goswift:apps:mobile:features:settings"))
-    implementation(project(":applications:goswift:apps:mobile:features:hydration"))
+    implementation(project(":features:health:hydration"))
     implementation(project(":applications:goswift:apps:mobile:features:input"))
 
     implementation(libs.androidx.health.connect.client)
