--- applications/goswift/apps/mobile/features/hydration/.gitignore
@@ -1,1 +1,0 @@
-/build
