--- applications/goswift/apps/mobile/features/hydration/build.gradle.kts
@@ -1,33 +1,0 @@
-plugins {
-    id("composetemplate.android.library")
-    id("composetemplate.android.library.compose")
-    id("composetemplate.android.hilt")
-}
-
-android {
-    namespace = "com.zoewave.probase.goswift.mobile.hydration"
-
-    defaultConfig {
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-}
-
-dependencies {
-    implementation(project(":core:ui"))
-    implementation(project(":core:util"))
-    implementation(project(":applications:goswift:model"))
-    implementation(project(":applications:goswift:data"))
-    implementation(project(":applications:goswift:features:main"))
-    implementation(libs.androidx.health.connect.client)
-    implementation(libs.androidx.compose.material3)
-    implementation(libs.androidx.compose.material.icons.extended)
-    implementation(libs.androidx.hilt.lifecycle.viewmodel.compose)
-    implementation(libs.androidx.lifecycle.runtime.compose)
-
-    // --- Testing ---
-    testImplementation(libs.junit)
-    testImplementation(libs.mockk)
-    testImplementation(libs.kotlinx.coroutines.test)
-    testImplementation(libs.truth)
-}
