--- applications/goswift/apps/mobile/features/hydration/proguard-rules.pro
@@ -1,5 +1,0 @@
-# Add project specific ProGuard rules here.
-# For more details, see http://developer.android.com/guide/developing/tools/proguard.html
-
-# Keep all classes in this package
--keep class com.zoewave.probase.goswift.mobile.hydration.** { *; }
